refactor: move policy classes to original package
