+++ spi/core-spi/src/main/java/org/eclipse/dataspaceconnector/spi/policy/AtomicConstraintFunction.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.spi.policy.evaluation;
+package org.eclipse.dataspaceconnector.spi.policy;
 
 import org.eclipse.dataspaceconnector.policy.model.Operator;
 import org.eclipse.dataspaceconnector.policy.model.Rule;
