+++ spi/core-spi/src/main/java/org/eclipse/dataspaceconnector/spi/policy/PolicyEngine.java
@@ -19,8 +19,6 @@ import org.eclipse.dataspaceconnector.policy.model.AtomicConstraint;
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.policy.model.Rule;
 import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgent;
-import org.eclipse.dataspaceconnector.spi.policy.evaluation.AtomicConstraintFunction;
-import org.eclipse.dataspaceconnector.spi.policy.evaluation.RuleFunction;
 import org.eclipse.dataspaceconnector.spi.result.Result;
 
 import java.util.Map;
